fixed error that wasPlayerOnline wasn't set corretly
also corrected saveing of playerAssignments.

the error occured at least in singlePlayer because giants passes to
addUser the static uuid SINGLEPLAYER_UUID and at onUserJoinGame the real
uuid is passed so we have a conflict. to fix that i take the uuid from
the user with the userId in addUser.